Cleanup: Move existing cases to example directory
Code structure improvements are not feasible to be carried out on all
existing cases.
The improvements will be done to a representative TNO-case which serves
as a starting point for future use.
The examples are kept and at a later point filtered to keep the ones which
show useful features (geographically merging inventories, combining tracers
from different inventories).