Move COSMODomain to grids-file, add InventoryGrid class
COSMODomain is now called COSMOGrid and can compute intersections
between it's gridcells and a inventory cell.

InventoryGrid is an abstract base class for inventory grids. To use
it for a new inventory, inherit from it and implement the virtual
methods (give the corners of the cell (i,j), give lat/lon ranges)

Updates to config:
	Fix bug where to cosmo-grid values were not updated when
	using offline emissions

	Use the TNOGrid class, so in theory the specifics of the
	TNO-grid can be read from the dataset and don't have to
	be given.

The code in this commit was not tested!